Add changes
Add VFX, fix code and problems with objects on game scene. Archived the folder with scenes because of to the size of the game scene. Change build. Рut a little order in the project files

+++ Assets/Source/UI/ResourceDisplay.cs
@@ -36,9 +36,9 @@ public class ResourceDisplay : MonoBehaviour
     private string FormatNumber(int num)
     {
         if (num >= 1000000)
-            return (num / 1000000).ToString() + "M";
+            return ((float)num / 1000000).ToString() + "M";
         else if (num >= 1000)
-            return (num / 1000).ToString() + "K";
+            return ((float)num / 1000).ToString() + "K";
         else
             return num.ToString();
     }
